Revert "Remove redundant hashes in bookmarks"
This reverts commit c85f87fea2f69684d52da393f50b005130339bf3.

+++ x64_dbg_dbg/breakpoint.cpp
@@ -48,7 +48,7 @@ int BpGetList(std::vector<BREAKPOINT>* List)
         }
     }
 
-    return (int)breakpoints.size();
+    return breakpoints.size();
 }
 
 bool BpNew(uint Address, bool Enable, bool Singleshot, short OldBytes, BP_TYPE Type, DWORD TitanType, const char* Name)
